Cat_Dog_Alexnet/backend/requirements.txt
fastapi
uvicorn
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
pillow
opencv-python
python-multipart